remove + defines

+++ sizes.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 main(){
-long long int i = 5;
+unsigned int i = 5;
 
 printf("size%lu\n",sizeof(i));
 }
